requirements.txt
--extra-index-url http://pypi.bars-open.ru/simple/
--trusted-host pypi.bars-open.ru
django==2.2
m3-objectpack==2.2.46
Django==2.2
pytz==2021.3
sqlparse==0.4.2
